requirements/requirements_new.txt
amqp==2.1.4
ansible==2.4.2.0
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.3
boto3==1.6.5
botocore==1.9.5
celery==4.1.1
certifi==2018.1.18
cffi==1.11.2
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.1.4
decorator==4.1.2
Django==2.0.6
django-auth-ldap==1.6.1
django-bootstrap3==10.0.1
django-celery-beat==1.1.1
django-filter==1.1.0
django-formtools==2.1
django-ranged-response==0.2.0
django-redis-cache==1.7.1
django-rest-swagger==2.2.0
django-simple-captcha==0.5.7
djangorestframework==3.8.2
djangorestframework-bulk==0.2.1
docutils==0.14
ecdsa==0.13
elasticsearch==6.1.1
enum-compat==0.0.2
ephem==3.7.6.0
eventlet==0.22.1
ForgeryPy==0.1
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jms-storage==0.0.17
kombu==4.2.1
ldap3==2.4
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
oss2==2.4.0
paramiko==2.4.0
passlib==1.7.1
Pillow==4.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
pyldap==2.4.45
PyNaCl==1.2.1
pyotp==2.2.6
python-dateutil==2.6.1
python-gssapi==0.6.4
python-ldap==3.1.0
pytz==2018.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
simplejson==3.13.2
six==1.11.0
sshpubkeys==2.2.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
